fix: reduced complexity caused by typescript definitions
