+++ src/app/constant/regex.ts
@@ -5,4 +5,4 @@ export const SUPPLIER_CODE = /^\d{6}$/;
 
 export const PASSWORD_STRENGTH = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[^]{8,}$/;
 
-export const ACCOUNT_REGEXP = /^((1\d{10})|(\w[-\w.+]*@([A-Za-z0-9][-A-Za-z0-9]+\.)+[A-Za-z]{2,14}))$/;
+export const ACCOUNT_REGEXP = /^((1[3456789]\d{9})|(\w[-\w.+]*@([A-Za-z0-9][-A-Za-z0-9]+\.)+[A-Za-z]{2,14}))$/;
